requirements/default.txt
boto3==1.7.4 \
    --hash=sha256:13ad5f64a247d655a27dca83274588e9d14cba61b38d3d4fd2b011b7197d88dd \
    --hash=sha256:a56b21efbc994580fc9cef454f0f949745c152326f939aed6609d1c47b2a0f8f
Django==2.0.4 \
    --hash=sha256:2d8b9eed8815f172a8e898678ae4289a5e9176bc08295676eff4228dd638ea61 \
    --hash=sha256:d81a1652963c81488e709729a80b510394050e312f386037f26b54912a3a10d0
python-dotenv==0.8.2 \
    --hash=sha256:4965ed170bf51c347a89820e8050655e9c25db3837db6602e906b6d850fad85c \
    --hash=sha256:509736185257111613009974e666568a1b031b028b61b500ef1ab4ee780089d5
click==6.7 \
    --hash=sha256:29f99fc6125fbc931b758dc053b3114e55c77a6e4c6c3a2674a2dc986016381d \
    --hash=sha256:f15516df478d5a56180fbf80e68f206010e6d160fc39fa508b65e035fd75130b
raven==6.7.0 \
    --hash=sha256:e4edf648829a64234800a10ed94ca08e0b38592f7449fa5e70931db62f5cd851 \
    --hash=sha256:f908e9b39f02580e7f822030d119ed3b2e8d32300a2fec6373e5827d588bbae7
dj-database-url==0.5.0 \
    --hash=sha256:4aeaeb1f573c74835b0686a2b46b85990571159ffc21aa57ecd4d1e1cb334163 \
    --hash=sha256:851785365761ebe4994a921b433062309eb882fedd318e1b0fcecc607ed02da9
psycopg2-binary==2.7.4 \
    --hash=sha256:02eb674e3d5810e19b4d5d00720b17130e182da1ba259dda608aaf33d787347d \
    --hash=sha256:3a14baeabcebd4662f12f4bff03e0574a2369a2e41baf829e6fb4a24c95cf88b \
    --hash=sha256:436a503eda41f6adb08f292f40a3784fce0a5f351b6ae7b19a911904db53af93 \
    --hash=sha256:465ff1d427ed42c31e456dbbd9edab3552be18a0edaef7450c5b3e6fee745052 \
    --hash=sha256:4a1a5ea2fa4b53191637b162873a82822d92a85a08beefe28296b8eb5cf2fea5 \
    --hash=sha256:4a4f23a08fbccbe40ecdb5384d807bcb469ea71dd87e6be2e80b036b8e6d47df \
    --hash=sha256:77a2fc622a1f2d08a707673c9be5769d521f03d867d305f172bb417fa7882754 \
    --hash=sha256:8014c06a9ed7b78ba81beff3ae71acd78c212390f8ed839e9ce22735880bd5b4 \
    --hash=sha256:83af04029bcb4b56c852e5876fef71340dcb465fa44fc99f80bac72e10fb0b74 \
    --hash=sha256:86c0d2587f56776f25d52cca8e275adf495c8e01933fbfc2ca23b124610ab761 \
    --hash=sha256:9305d7cbc802aaefac5c75a3df725f2654797369f32b18d4d0adb382dfab6c09 \
    --hash=sha256:9b5ddbed85ec73293695d7116589d956ef0dd3fcf7bf3b2a3bc1e8e54c1d543a \
    --hash=sha256:a3d2cc0cb0b988dbfd0d11f7fac34058b25a6ce533ed5b8e88d6cb315e77d54a \
    --hash=sha256:ab1db8f3e96570d9f7ebc45133ce2574804b2280499baade178e163d022107b5 \
    --hash=sha256:b039f51bca1ddd70234cc3f84f94f42ad43861b931bdfb497f887c60c39a6565 \
    --hash=sha256:b287ddf4cafcfb632974907d1e7862119e36bb758228bdb07dd247553e4cdfc0 \
    --hash=sha256:b6b2b26590304d97ef2af28d153ee99ace6fe0806934f4618edfc87216c77f91 \
    --hash=sha256:c4c6004d410c77bfa5389ae9485498ce32805447a67afbfe8db0d247a5c88fa1 \
    --hash=sha256:c606bff0978ee4858d86d40f6b6ab0c4cac4474f627bd054683dc03a4fc1a366 \
    --hash=sha256:c8220c521a408b41c4f14036004a621ed0d965941286b978cd2ea2623fabd755 \
    --hash=sha256:cb07184a4bfad304831f0a88b1c13fbd8cf9fcdf1f11e71c477dd6d7b1b078a0 \
    --hash=sha256:cf3911fba0c47fc1313b5783183cda301032b14637a0b7a336766ae46998c7ee \
    --hash=sha256:d0972f062c73956332e9681dfdb133168618f0abfecc96e89f0205ac89cd454b \
    --hash=sha256:d1dd3eb8edd354083f5d27b968c5a17854c41347ba5a480b520be85ec1a8495c \
    --hash=sha256:d51c7ed810fce1e50464088c37cc8da05534de8afb12a732500827ebcc480081 \
    --hash=sha256:d8940b5104588d6313315e037f0f5ed68d2e5f62ccc1c429d3cff11d2ba6de3f \
    --hash=sha256:de4f88f823037a71ea5ef3c1041d96b8a68d73343133edda684fd42f575bd9d7
requests==2.18.4 \
    --hash=sha256:6a1b267aa90cac58ac3a765d067950e7dbbf75b1da07e895d1f594193a40a38b \
    --hash=sha256:9c443e7324ba5b85070c4a818ade28bfabedf16ea10206da1132edaa6dda237e
Jinja2==2.10 \
    --hash=sha256:74c935a1b8bb9a3947c50a54766a969d4846290e1e788ea44c1392163723c3bd \
    --hash=sha256:f84be1bb0040caca4cea721fcbbbbd61f9be9464ca236387158b0feea01914a4
datadog==0.20.0 \
    --hash=sha256:7bb4af836d7422a6138f983b81c16acd56c2d608913982602856cc273ae74768
dockerflow==2018.4.0 \
    --hash=sha256:2ea52a904abfda3430ff4f1effc164863b30d2b69f7ecbf92dd672860b0ec423 \
    --hash=sha256:388d02c557968e6957140f7b82f669eac70adf5f570bc7705aa749d220a2e535
pytest==3.5.1 \
    --hash=sha256:54713b26c97538db6ff0703a12b19aeaeb60b5e599de542e7fca0ec83b9038e8 \
    --hash=sha256:829230122facf05a5f81a6d4dfe6454a04978ea3746853b2b84567ecf8e5c526
pytest-django==3.2.1 \
    --hash=sha256:534505e0261cc566279032d9d887f844235342806fd63a6925689670fa1b29d7 \
    --hash=sha256:7501942093db2250a32a4e36826edfc542347bb9b26c78ed0649cdcfd49e5789
pytest-timeout==1.2.1 \
    --hash=sha256:142739bc1f0687fccb0f013d36c5bbebd0b14b546404e79f4158010e81e5013b \
    --hash=sha256:68b7d264633d5d33ee6b14ce3a7f7d05f8fd9d2f6ae594283221ec021736b7cd
pytest-mock==1.10.0 \
    --hash=sha256:53801e621223d34724926a5c98bd90e8e417ce35264365d39d6c896388dcc928 \
    --hash=sha256:d89a8209d722b8307b5e351496830d5cc5e192336003a485443ae9adeb7dd4c0
factory_boy==2.11.1 \
    --hash=sha256:6f25cc4761ac109efd503f096e2ad99421b1159f01a29dbb917359dcd68e08ca \
    --hash=sha256:d552cb872b310ae78bd7429bf318e42e1e903b1a109e899a523293dfa762ea4f
